services/visionary-agent/requirements.txt
dspy-ai>=2.5.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-arango>=7.5.0
anthropic>=0.40.0
httpx>=0.25.0
python-dotenv>=1.0.0
